Issue 198:  mjpegstream not stoping
Improving a bit reaction on thread aborting for JPEGStream and MJPEGStream classes, so they return faster when ThreadAbortException exception is received. The rest of the issue was fixed before when improving HTTP time outs handling. Now Stop() will not return immediately also, but still will return much faster than it was in 2.1.5, where it was kind of blocking caller thread if camera got unplugged.